Convert to Grails 3 plugin
