+++ .gitignore
@@ -1,10 +1,14 @@
-.idea/
 *.iml
+*.iws
 *.log
-/target
-/docs
-grails-spring-security-cas-*.zip
-grails-spring-security-cas-*.zip.sha1
-plugin.xml
+*.sublime-workspace
+.DS_Store
+.gradle
+.idea
+.nb-gradle
 .settings
-web-app
+bin
+build
+cobertura.ser
+target
+target-eclipse
