+++ bintrayPublishing.gradle
@@ -1,0 +1,32 @@
+bintray {
+    user = System.getenv('BINTRAY_USER') ?: project.hasProperty('bintrayUser') ? project.bintrayUser : ''
+    key = System.getenv('BINTRAY_KEY') ?: project.hasProperty('bintrayKey') ? project.bintrayKey : ''
+    publications = ['maven']
+    publish = true
+    pkg {
+        repo = 'plugins'
+        userOrg = 'grails'
+        name = "org.grails.plugins:$project.name"
+        desc = "Grails $project.name plugin"
+        websiteUrl = "http://grails.org/plugin/$project.name"
+        issueTrackerUrl = 'https://github.com/grails-plugins/grails-spring-security-cas/issues'
+        vcsUrl = 'https://github.com/grails-plugins/grails-spring-security-cas'
+        licenses = ['Apache-2.0']
+        publicDownloadNumbers = true
+        version {
+            attributes = ['grails-plugin': "org.grails.plugins:$project.name"]
+            name = project.version
+            gpg {
+                sign = false
+                passphrase = System.getenv('SIGNING_PASSPHRASE') ?: project.hasProperty('signingPassphrase') ? project.signingPassphrase : ''
+            }
+            mavenCentralSync {
+                sync = false
+                def ossUser = System.getenv('SONATYPE_USERNAME') ?: project.hasProperty('sonatypeOssUsername') ? project.sonatypeOssUsername : ''
+                def ossPass = System.getenv('SONATYPE_PASSWORD') ?: project.hasProperty('sonatypeOssPassword') ? project.sonatypeOssPassword : ''
+                user = ossUser
+                password = ossPass
+            }
+        }
+    }
+}
