--- build.xml
@@ -1,72 +1,0 @@
-<project name='spring-security-cas' default='package'>
-
-	<!-- Properties -->
-
-	<property file='application.properties' />
-
-	<property environment='env'/>
-
-	<!-- optional properties file for developer overrides -->
-	<property file='build.properties' />
-
-	<property name='grails.home' value='${env.GRAILS_HOME}' />
-
-	<condition property='grails' value='grails.bat'>
-		<os family='windows'/>
-	</condition>
-	<property name='grails' value='grails' />
-
-	<!-- Macrodefs -->
-
-	<macrodef name='grails'>
-		<attribute name='action' />
-		<attribute name='environment' default='dev' />
-		<element name='args' optional='true' />
-		<sequential>
-			<exec executable='${grails}' failonerror='true'>
-				<args />
-				<arg value='@{environment}'/>
-				<arg value='@{action}'/>
-			</exec>
-		</sequential>
-	</macrodef>
-
-	<!-- Targets -->
-
-	<target name='init'>
-		<mkdir dir='grails-app/conf/spring'/>
-	</target>
-
-	<target name='clean' description='Cleans a Grails application' depends='init'>
-		<grails action='clean' />
-		<delete><fileset dir='.' includes='*.log*' /></delete>
-	</target>
-
-	<target name='test' description='Run unit tests' depends='clean'>
-		<grails action='test-app' environment='test' />
-	</target>
-
-	<target name='package' description='Package the plugin'
-	        depends='test, doPackage, post-package-cleanup'/>
-
-	<target name='doPackage'>
-		<grails action='package-plugin' />
-	</target>
-
-	<target name='post-package-cleanup'>
-		<delete dir='grails-app/conf/hibernate'/>
-		<delete dir='grails-app/conf/spring'/>
-		<delete dir='grails-app/controllers'/>
-		<delete dir='grails-app/domain'/>
-		<delete dir='grails-app/i18n'/>
-		<delete dir='grails-app/services'/>
-		<delete dir='grails-app/taglib'/>
-		<delete dir='grails-app/utils'/>
-		<delete dir='grails-app/views'/>
-		<delete dir='lib'/>
-		<delete dir='src/groovy'/>
-		<delete dir='test'/>
-		<delete dir='web-app'/>
-	</target>
-
-</project>
