--- grails-app/conf/Config.groovy
@@ -1,10 +1,0 @@
-// not included in plugin zip
-
-// dummy values to get package to run
-grails.plugin.springsecurity.cas.loginUri = '/login'
-grails.plugin.springsecurity.cas.serviceUrl = 'http://localhost:8080/castest/j_spring_cas_security_check'
-grails.plugin.springsecurity.cas.serverUrlPrefix = 'http://localhost:9090/cas-server-webapp-3.4.2'
-
-grails.doc.authors = 'Burt Beckwith'
-grails.doc.license = 'Apache License 2.0'
-grails.doc.title = 'Spring Security CAS Plugin'
