+++ grails-app/conf/DefaultCasSecurityConfig.groovy
@@ -1,4 +1,4 @@
-/* Copyright 2006-2013 the original author or authors.
+/* Copyright 2006-2015 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
