+++ grails-app/conf/application.yml
@@ -1,0 +1,12 @@
+---
+grails:
+    profile: web-plugin
+    doc:
+        authors: 'Burt Beckwith'
+        license: 'Apache License 2.0'
+        title: 'Spring Security CAS Plugin'
+info:
+    app:
+        name: '@info.app.name@'
+        version: '@info.app.version@'
+        grailsVersion: '@info.app.grailsVersion@'
