+++ release.sh
@@ -1,10 +1,9 @@
-rm -rf target/release
-mkdir target/release
-cd target/release
+rm -rf build/release
+mkdir -p build/release
+cd build/release
 git clone git@github.com:grails-plugins/grails-spring-security-cas.git
 cd grails-spring-security-cas
 grails clean
 grails compile
 
-#grails publish-plugin --snapshot --stacktrace
-grails publish-plugin --stacktrace
+gradle bintrayUpload --stacktrace
